# career-quest-api/public/api/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"securitySchemes": {}
	},
	"info": {
		"title": "career-quest-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/v1/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"user": {
											"properties": {
												"ResetPasswordToken": {
													"type": "string"
												},
												"UpdatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"CreatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"ReferredById": {
													"type": "number",
													"format": "double"
												},
												"ReferralCode": {
													"type": "string"
												},
												"Gender": {
													"type": "string"
												},
												"Phone": {
													"type": "string"
												},
												"Address": {
													"type": "string"
												},
												"Password": {
													"type": "string"
												},
												"Email": {
													"type": "string"
												},
												"Username": {
													"type": "string"
												},
												"UserId": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"ResetPasswordToken",
												"UpdatedAt",
												"CreatedAt",
												"ReferredById",
												"ReferralCode",
												"Gender",
												"Phone",
												"Address",
												"Password",
												"Email",
												"Username",
												"UserId"
											],
											"type": "object"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"user",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "referralCode",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/profile/{userId}": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"ResetPasswordToken": {
											"type": "string"
										},
										"UpdatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"CreatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"ReferredById": {
											"type": "number",
											"format": "double"
										},
										"ReferralCode": {
											"type": "string"
										},
										"Gender": {
											"type": "string"
										},
										"Phone": {
											"type": "string"
										},
										"Address": {
											"type": "string"
										},
										"Password": {
											"type": "string"
										},
										"Email": {
											"type": "string"
										},
										"Username": {
											"type": "string"
										},
										"UserId": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"ResetPasswordToken",
										"UpdatedAt",
										"CreatedAt",
										"ReferredById",
										"ReferralCode",
										"Gender",
										"Phone",
										"Address",
										"Password",
										"Email",
										"Username",
										"UserId"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/auth/status": {
			"get": {
				"operationId": "CheckLoginStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"user": {
											"properties": {
												"ResetPasswordToken": {
													"type": "string"
												},
												"UpdatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"CreatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"ReferredById": {
													"type": "number",
													"format": "double"
												},
												"ReferralCode": {
													"type": "string"
												},
												"Gender": {
													"type": "string"
												},
												"Phone": {
													"type": "string"
												},
												"Address": {
													"type": "string"
												},
												"Password": {
													"type": "string"
												},
												"Email": {
													"type": "string"
												},
												"Username": {
													"type": "string"
												},
												"UserId": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"ResetPasswordToken",
												"UpdatedAt",
												"CreatedAt",
												"ReferredById",
												"ReferralCode",
												"Gender",
												"Phone",
												"Address",
												"Password",
												"Email",
												"Username",
												"UserId"
											],
											"type": "object"
										},
										"token": {
											"type": "string"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"user",
										"token",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/auth/referrals/{userId}": {
			"get": {
				"operationId": "GetReferrals",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"referredUsers": {
											"items": {
												"properties": {
													"CreatedAt": {
														"type": "string",
														"format": "date-time"
													},
													"Email": {
														"type": "string"
													},
													"Username": {
														"type": "string"
													},
													"UserId": {
														"type": "number",
														"format": "double"
													}
												},
												"required": [
													"CreatedAt",
													"Email",
													"Username",
													"UserId"
												],
												"type": "object"
											},
											"type": "array"
										},
										"referredBy": {
											"properties": {
												"Email": {
													"type": "string"
												},
												"Username": {
													"type": "string"
												},
												"UserId": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"Email",
												"Username",
												"UserId"
											],
											"type": "object"
										}
									},
									"required": [
										"referredUsers",
										"referredBy"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}